+++ tests/Sabre/CalDAV/Notifications/NodeTest.php
@@ -15,7 +15,7 @@ class NodeTest extends \PHPUnit_Framework_TestCase {
 
         $this->systemStatus = new Notification\SystemStatus(1,'"1"');
 
-        $this->caldavBackend = new CalDAV\Backend\Mock(array(),array(), array(
+        $this->caldavBackend = new CalDAV\Backend\MockSharing(array(),array(), array(
             'principals/user1' => array(
                 $this->systemStatus
             )